[Chore] #26 - 필요 없는 코드 삭제
